backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.9"
Django = "^3.1.3"
bandit = "^1.6.2"
djangorestframework = "^3.12.2"
drf-yasg = "^1.20.0"
django-password-reset = "^2.0"
gunicorn = "^20.0.4"
django-heroku = "0.3.0"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
flake8 = "^3.8.4"
pytest = "^6.1.2"
pytest-django = "^4.1.0"
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
